+++ internal/quickstart/container.go
@@ -7,7 +7,7 @@ import (
 	"github.com/charmbracelet/bubbles/key"
 	"github.com/charmbracelet/bubbles/spinner"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
+	"github.com/muesli/reflow/wordwrap"
 
 	"ldcli/internal/flags"
 )
@@ -35,10 +35,10 @@ type ContainerModel struct {
 	err          error
 	flagKey      string
 	flagsClient  flags.Client
-	quitMsg      string // TODO: set this?
 	quitting     bool
 	sdk          sdkDetail
 	totalSteps   int
+	width        int
 }
 
 func NewContainerModel(flagsClient flags.Client, accessToken string, baseUri string) tea.Model {
@@ -59,16 +59,34 @@ func (m ContainerModel) Init() tea.Cmd {
 func (m ContainerModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	var cmd tea.Cmd
 	switch msg := msg.(type) {
+	case tea.WindowSizeMsg:
+		m.width = msg.Width
 	case tea.KeyMsg:
 		switch {
-		case key.Matches(msg, keys.Quit):
+		case key.Matches(msg, pressableKeys.Quit):
 			m.quitting = true
 			cmd = tea.Quit
-		case key.Matches(msg, keys.Back):
-			// if showing SDK instructions, let the user go back to choose a different SDK
-			if m.currentStep == stepShowSDKInstructions {
+		case key.Matches(msg, pressableKeys.Back):
+			switch m.currentStep {
+			case stepCreateFlag:
+				// can't go back
+			case stepChooseSDK:
+				m.currentStep -= 1
+				m.currentModel = NewCreateFlagModel(m.flagsClient, m.accessToken, m.baseUri)
+			case stepShowSDKInstructions:
 				m.currentStep -= 1
 				m.currentModel = NewChooseSDKModel(m.sdk.index)
+				cmd = m.currentModel.Init()
+			case stepToggleFlag:
+				m.currentStep -= 1
+				m.currentModel = NewShowSDKInstructionsModel(
+					m.accessToken,
+					m.baseUri,
+					m.sdk.canonicalName,
+					m.sdk.displayName,
+					m.sdk.url,
+					m.flagKey,
+				)
 				cmd = m.currentModel.Init()
 			}
 		default:
@@ -93,7 +111,7 @@ func (m ContainerModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		m.currentStep += 1
 		m.err = nil
 	case errMsg:
-		m.err = msg.err
+		m.currentModel, cmd = m.currentModel.Update(msg)
 	case noInstructionsMsg:
 		// skip the ShowSDKInstructionsModel and move along to toggling the flag
 		m.currentModel = NewToggleFlagModel(
@@ -126,55 +144,9 @@ func (m ContainerModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 func (m ContainerModel) View() string {
 	out := fmt.Sprintf("\nStep %d of %d\n"+m.currentModel.View(), m.currentStep, m.totalSteps)
 
-	if m.err != nil {
-		if m.quitting {
-			out := m.quitMsg + "\n\n"
-			out += m.err.Error()
-
-			return lipgloss.
-				NewStyle().
-				SetString(out).
-				Render() + "\n"
-		}
-
-		// show error and stay on the same step
-		out += "\n" + lipgloss.
-			NewStyle().
-			SetString(m.err.Error()).
-			Render() + "\n"
-
-		return out
-	}
-
 	if m.quitting {
 		return ""
 	}
 
-	return out
-}
-
-type keyMap struct {
-	Back  key.Binding
-	Enter key.Binding
-	Quit  key.Binding
-	Tab   key.Binding
-}
-
-var keys = keyMap{
-	Back: key.NewBinding(
-		key.WithKeys("esc"),
-		key.WithHelp("esc", "go back"),
-	),
-	Enter: key.NewBinding(
-		key.WithKeys("enter"),
-		key.WithHelp("enter", "select"),
-	),
-	Quit: key.NewBinding(
-		key.WithKeys("ctrl+c"),
-		key.WithHelp("q", "quit"),
-	),
-	Tab: key.NewBinding(
-		key.WithKeys("tab"),
-		key.WithHelp("tab", "toggle"),
-	),
+	return wordwrap.String(out, m.width)
 }